# api/openapi.yaml
openapi: 3.0.0
info:
  title: EduDocsHub API
  version: 1.0.0

#servers:
#- url: https://EdudocsHub.com/api
...
paths:
...
  /admin/courses/{courseid}:
    put:
      tags:
      - admin
      summary: Admin Course Editing
      description: "We send the new characteristics of the course and receive a \n\
        confirmation message as a response.\n"
      operationId: adminCourseEditing
      parameters:
      - name: courseid
        in: path
        description: |
          The ID of the course that the admin is going to edit.
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Course'
      responses:
        "200":
          description: Successfully changed the course attributes message.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: Admin
    delete:
      tags:
      - admin
      summary: Admin file Deletion
      description: Admin removes a file
      operationId: adminRemoveCourse
      parameters:
      - name: courseid
        in: path
        description: The courseid in order to get the course object.
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: Successfully deleted course.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: Admin
components:
  schemas:
    Message:
      type: object
      properties:
        text:
          type: string
      example:
        text: text
    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
...
    Course:
      type: object
      properties:
        semester:
          type: string
        school:
          type: string
        university:
          type: string
        title:
          type: string
        courseid:
          type: integer
      example:
        school: school
        university: university
        semester: semester
        title: title
        courseid: 0
...